+++ HISTORY.rst
@@ -2,7 +2,10 @@
 Important History
 ==================
 
--1.3.4 bug fixes for the dutch language
+- 1.4.4 added support for papago, added opt params, fixed deepl free api
+- 1.4.3 added support deepl free api
+- 1.4.2 added proxy support
+- 1.3.4 bug fixes for the dutch language
 - 1.3.3 fixed bug in google translate
 
 - 1.2.5: added support for the DeepL translator